requirements.txt
python-pptx
pandas